leadgen-api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
celery>=5.3.6
redis>=5.0.1
playwright>=1.40.0
aiosmtplib>=3.0.1
dnspython>=2.4.2
loguru>=0.7.2
tenacity>=8.2.3
pydantic>=2.4.2
pydantic-settings>=2.1.0
requests>=2.31.0
httpx>=0.25.2
